Término da implementação das entidades, seguindo o modelo de filme; Acrescento da tela e da coluna de listagem de ingressos por usuário, aonde o funcionário do cinema poderá ver todos os ingressos consumidos por aquele usuário, em específico

+++ src/main/java/br/com/senai/cinema/resources/UsuarioResource.java
@@ -3,9 +3,9 @@ package br.com.senai.cinema.resources;
 import java.util.List;
 
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,16 +13,13 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
 
-import com.google.gson.Gson;
-
+import br.com.senai.cinema.models.Ingresso;
 import br.com.senai.cinema.models.Usuario;
 import br.com.senai.cinema.services.UsuarioService;
 import javassist.tools.rmi.ObjectNotFoundException;
 
 @Controller
-//@RestController
 @RequestMapping(value="/usuario")
 public class UsuarioResource {
 		
@@ -38,45 +35,36 @@ public class UsuarioResource {
 		return "usuario";
 	}
 	
-//	@GetMapping("/listagem")
-//	public List<Usuario> usuarios() {
-//		return  usuarioService.findAll();
-//	}
-	
-	@GetMapping("/listagem/{id}")
-	public Usuario getUsuario(@PathVariable(value="id") Integer id) throws ObjectNotFoundException {
-		return  usuarioService.findById(id);
+	@PutMapping("/update/{id}")
+	public ResponseEntity<?> update(@PathVariable Integer id, @RequestBody Usuario usuario) throws ObjectNotFoundException {
+		Usuario objUsuario = usuarioService.findById(id);
+		objUsuario.setNome(usuario.getNome());
+		objUsuario.setEmail(usuario.getEmail());
+		objUsuario.setLogin(usuario.getLogin());
+		objUsuario.setSenha(usuario.getSenha());
+		usuarioService.update(objUsuario);
+		
+		return ResponseEntity.ok().body("atualizado");
 	}
 	
-	@PutMapping("/update-usuario")
-	public @ResponseBody String updateReturningJson(HttpServletRequest request, HttpServletResponse response) throws ObjectNotFoundException {
-
-		Integer idUsuario = Integer.parseInt(request.getParameter("idUsuario"));
+	@DeleteMapping(value = "/delete/{id}")
+	public ResponseEntity<?> delete(@PathVariable("id") String id) throws NumberFormatException, ObjectNotFoundException {
+		Usuario usuario = usuarioService.findById(Integer.parseInt(id));
+		if (usuario != null) {
+			usuarioService.delete(usuario);
+		}
+		return ResponseEntity.ok().body("excluido");
+	}
+	
+	@GetMapping("/ingressos/{id}")
+	public String listAllIngressosByUsuario(@PathVariable Integer idUsuario, HttpServletRequest request) throws ObjectNotFoundException {
+		
 		Usuario usuario = usuarioService.findById(idUsuario);
-		request.setAttribute("usuario", usuario);
-		
-		Gson gson = new Gson();
-		String json = gson.toJson(usuario);
-		
-		return json;
-	}
-	
-	@PutMapping("/listagem") // @RequestBody = obj vem no corpo da requisição
-	public Usuario update(@RequestBody Usuario usuario) {
-		return usuarioService.update(usuario);
-	}
-	
-	@DeleteMapping("/listagem") // @RequestBody = obj vem no corpo da requisição
-	public void delete(@RequestBody Usuario usuario) {
-		usuarioService.delete(usuario);
-	}
-	
-	@DeleteMapping("/delete-usuario") // @RequestBody = obj vem no corpo da requisição
-	public String deleteById(HttpServletRequest request) {
-		String idUsuario = request.getParameter("idUsuario");
-		System.out.println("Id Usuario" + idUsuario);
-		usuarioService.deleteById(Integer.parseInt(idUsuario));
-		return "usuario";
+
+		List<Ingresso> ingressos = usuarioService.listAllIngressosByUsuario(usuario);
+
+		request.setAttribute("ingressos", ingressos);
+		return "listagem-de-ingressos-por-usuario";
 	}
 	
 }
